requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
mysql-connector-python==8.0.31
mysqlclient==2.1.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
